+++ sale_multi_operating_unit/models/__init__.py
@@ -1,2 +1,6 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
-from . import sale_order
+from . import (
+    sale_order_quote_line,
+    sale_order_quote,
+    sale_order,
+)
