+++ sale_multi_operating_unit/views/sale_order_view.xml
@@ -13,7 +13,8 @@
                     <field name="quote_ids" context="{'active_id': active_id}">
                         <tree>
                             <field name="operating_unit_id"/>
-                            <field name="lead_id"/>
+                            <field name="lead_id" attrs="{'invisible': [('lead_id', '=', False)]}"/>
+                            <button name="action_send" type="object" string="Send" class="oe_highlight" attrs="{'invisible': [('lead_id', '!=', False)]}"/>
                         </tree>
                         <form string="Sale Multiple Operating Unit">
                             <header>
@@ -30,10 +31,10 @@
                                     <field name="operating_unit_id"/>
                                 </group>
                                 <group>
-                                    <field name="lead_id" force_save="1"/>
+                                    <field name="lead_id" force_save="1" attrs="{'invisible': [('lead_id', '=', False)]}"/>
                                 </group>
                             </group>
-                            <field name="line_ids" widget="section_and_note_one2many" mode="tree,kanban">
+                            <field name="line_ids">
                                 <form>
                                     <group>
                                         <field name="product_id"/>
@@ -45,10 +46,6 @@
                                     </group>
                                 </form>
                                 <tree editable="bottom">
-                                    <control>
-                                        <create string="Add a product"/>
-                                        <create string="Add Description"/>
-                                    </control>
                                     <field name="product_id"/>
                                     <field name="name"/>
                                     <field name="uom_id" groups="uom.group_uom"/>
